Day 05 Python: Oh no.
The natural solution to the given problem was to create a graph of the
rules and then do a topological sort to construct a total ordering on
the graph. Then, verifying that the pages are in order would be easy!
This works well on the test input, but it seems like the actual input
graph contains cycles, rendering a topological sort impossible.

Instead, I am going to have to rewrite the logic to ditch the
topological sort in favor of the more fuzzy "one node comes after
another node if the former is reachable from the latter." This will be
much slower, since I will have to walk the graph for every node.

All this is to say that I am committing this as a checkpoint before I
basically rewrite the whole thing.